Examples changes (#4865)
Various improvements to ordering and labelling of examples in the docs
site, based on a conversation I had in the discord. Now, useful and
important examples are being surfaced more optimally in the docs site.

This PR makes a few changes:

- Changes the title of the code x-ray example
- Adds tags for the code x-ray example
- References the code x-ray example from the custom ui example
- Adds priority fields to some of the examples
- Adds a default priority of 5 when sorting examples

For the future we might consider a lint rule to ensure all examples have
tags and priority set.

### Change type

- [ ] `bugfix`
- [x] `improvement`
- [ ] `feature`
- [ ] `api`
- [ ] `other`

### Test plan

### Release notes

- Improved examples naming, tags and priority

---------

Co-authored-by: Steve Ruiz <steveruizok@gmail.com>