Simplify e2e test running. (#4957)
Not sure why I didn't think of this before 🤦‍♂️

Simplify how we run e2e tests. Instead of having to run zero cache on
our own, we now completely rely on playwright. The way it works is:
- playwright checks port 3000 to see if the server is already running.
if it is not it will run `yarn dev-app` from the root folder
- this starts the dev scripts inside `apps/dotcom`
- this will start zero cache and will wake up docker and postgress
- `yarn dev` for `apps/dotcom/client` now waits for docker to be healthy
before proceeding. this means that playwright will only get a valid
response on port 3000 when docker is up. it also means it will clean all
the background processes after it finishes and correctly report the
status.

### Change type

- [ ] `bugfix`
- [x] `improvement`
- [ ] `feature`
- [ ] `api`
- [ ] `other`

### Release notes

- Simplify botcom e2e logic.