[botcom] simplify replicator dispatch logic (#4965)
This stuff was a little bit of a mess. I made it synchronous again and
hopefully caught an extra edge case or two.

- added a `isFileOwner` field to file_state, populated by triggers on
both the file table and the file_state table.
- used `isFileOwner` to avoid checking ownership while dispatching
events in the replicator
- simplified the event dispatching logic by moving it out into separate
methods for each table, so it's easier to follow what's happening.
- fixed the 'file deleting isn't working in prod' bug, that was to do
with r2 complaining about things it doesn't complain about in dev mode.


### Change type

- [ ] `bugfix`
- [ ] `improvement`
- [ ] `feature`
- [ ] `api`
- [x] `other`