[botcom] allow creating rooms quickly (#4990)
Before this PR, if you clicked the new room button a bunch of times
quickly, it would trigger a race condition where the room would end up
never actually being created and added to R2, which would result in
`404`s being returned from the server when you went to look at the rooms
you created later. This PR fixes that by having a fallback to check that
a file has been added to the DB but not R2 yet.

### Change type

- [x] `other`