Make sure notes snap to grid after position is updated
The notes are updated to be centered around the creation point. However,
the possible snapping to the grid was done before they were moved. This
meant that if the note width or height was not a multiple of the grid
size, they would not be snapped to the grid. This could occur e.g. if
dynamic size was on.