Fix an issue with drag and dropping the files (#5013)
We got an error saying there was a bad request when drag and dropping
files. This is because we wanted to create a file state as the file
owner, but the file did not exist yet (which is not allowed, we only
allow that for guest files). So we now also create a file together with
the file state.


![image](https://github.com/user-attachments/assets/2bde9622-e511-4f84-ad08-b1d6dd54e3cc)

I also noticed that the `createFile` function does not need to be async
and also extracted document names from the files if we have them.

### Change type

- [x] `bugfix`
- [ ] `improvement`
- [ ] `feature`
- [ ] `api`
- [ ] `other`

### Test plan

1. Drag some files to the sidebar.
2. They should correctly appear (with the correct names if they have
time) and you should not see any rejected mutation toasts (like the one
above).