Allow renaming files when pressing enter (#5019)
This allows renaming active files when the sidebar has focus and when
you press enter. When the file link becomes active it focuses itself and
when focused it enters rename mode if we press enter.

This also prevents creating a new file if we first click the create file
button and then press enter immediately. The focus changes to the new
file and we will start renaming it.

### Change type

- [ ] `bugfix`
- [x] `improvement`
- [ ] `feature`
- [ ] `api`
- [ ] `other`

### Test plan

1. Click create file button.
2. Press enter. You should not see another file getting created.
Instead, you should be in the rename mode for the newly added file.

---------

Co-authored-by: David Sheldrick <d.j.sheldrick@gmail.com>