Create a utility type for making undefined properties optional  (#5055)
Adapted from #5011.

This extracts a type used in the object validation for making all
properties of an object that accepts undefined optional. This is useful
when you have an object of validators (e.g. props for a shape) and want
to create the typescript type for an object that will be accepted by it.
With this, any of the properties that has a validator with `.optional()`
will be optional in that object.

### Change type

- [ ] `bugfix`
- [ ] `improvement`
- [ ] `feature`
- [ ] `api`
- [x] `other`

### Test plan

- [ ] Unit tests
- [ ] End to end tests

### Release notes

- Expose a utility type for making undefined properties optional

---------

Co-authored-by: Trygve Aaberge <trygve.aaberge@autodesk.com>