error page: wordsmith copy again (#4825)
continuing conversation from:
https://github.com/tldraw/tldraw/pull/4790#discussion_r1819122178
and i guess rehashing out what i thought we had settled in
https://github.com/tldraw/tldraw/pull/4782 :)

i'm gonna pull in @steveruizok here to mediate because i think we've
changed this back-and-forth enough times! let's talk about it on this PR
and bring forth our points.

- i understand that currently either things are fully private or fully
public but soon it will encompass invited users. _and_ we do have
`NOT_AUTHENTICATED` and `FORBIDDEN` distinctions for good reason: they
help guide the user to know whether the error is temporary or permanent
(i'm not logged in vs. i really don't have access even though i'm logged
in)
- furthermore, something labeled as a "private file", when somebody
wants to ask for access, is off-putting for a human to see. it feels
intrusive to ask for access to something that is explicitly called
private.
- finally, i do want to callout that we should avoid any paradigm that
publicly distinguishes to a non-authenticated/non-invited user that a
file is private vs. invite-only. this can be a slight leak of security
as it tells a snooping person what type of file is behind the access
controls. we should keep it intentionally a little obscure to avoid
that.

### Change type

- [ ] `bugfix`
- [ ] `improvement`
- [ ] `feature`
- [ ] `api`
- [x] `other`

---------

Co-authored-by: Steve Ruiz <steveruizok@gmail.com>