[botcom] move sidebar toggles in to sidebar layout so they're always visible (#5099)
Previously the sidebar toggle button would disappear when you see an
error page because it was only rendered as a child of the editor's
MenuPanel.

This was mitigated on desktop because we forced the sidebar open on
error pages, but that doesn't work on mobile because the sidebar is a
modal drawer.

So here I've just moved the toggle button to be rendered by the sidebar
layout component, and added a spacer div where it would be in the
MenuPanel.

this should be a pixel perfect match for what was there before, but it
obviously makes the menu panel code a tiny bit more fragile. but since
we see it all the time it's unlikely we'll introduce style regressions
by accident.

### Change type

- [ ] `bugfix`
- [ ] `improvement`
- [ ] `feature`
- [ ] `api`
- [x] `other`

---------

Co-authored-by: Steve Ruiz <steveruizok@gmail.com>