fix stale closure in InnerShape (#5117)
fixes #5047

the problem was we weren't destroying the shape effect schedulers when
the editor changed, because they were only keyed by the shape type name,
which obviously does not change when the editor is reinstantiated.

this augments useStateTracking to accept a deps array, and uses the
identity of the shape util as part of the deps array.

### Change type

- [x] `bugfix`