[botcom] Support legacy routes (#5123)
This PR adds (stubs) support for legacy routes in botcom. Those are:
- shared room
- shared room (readonly)
- shared room (readonly, old url)
- snapshot
- history
- history snapshot
- touchscreen sidebar (maybe)

I think it's better for us to treat these as "new routes" within botcom
so that we can develop out the ways that these routes should be handled.
Many routes need new UI and interactions, while the more rare internal
routes can be mostly left as they are.

We don't want to break shared rooms etc for people, however we do want
to have the prompts to sign in, etc, present on these rooms, similar to
what we display to signed out users for files. And for signed in users,
we should have a way to "claim" or "slurp" the file into your account's
files.

Some notes:

- titles appear in title bar for legacy routes
- legacy routes support copy tab in share menu


### Change type

- [x] `other`