Fix relative CSS import rules failing to be fetched
When getting fonts for an export, all stylesheets on the page are
traversed. If any CSS import rules are found, they are downloaded.
However this just directly used the href of the import rule, which makes
it relative to the document. Import rule URLs should however be relative
to the style sheet they are included in. We therefore have to construct
an absolute URL by using the style sheet as the base. If the import rule
doesn't have a parent style sheet it means it's an inline style sheet
and we should fall back to using the baseURI as the base.

Fixes #4939

Co-authored-by: Torkil Aamodt <torkil.aamodt@autodesk.com>