Gist embed: restrict the url to a more strict format (#5170)
A user reported that there was a slight risk in allowing Github Gist's
as they were currently being added to the canvas — basically, the JSONP
option on Github can be used to pass in arbitrary Javascript which can
look quite dodgy. But, in the end the Gist is iframed and thus,
sandboxed.

Nonetheless @MitjaBezensek and I worked to come up with a more
restrictive URL filter which should mitigate the issue.

Further reading:
-
https://github.com/bhaveshk90/Content-Security-Policy-CSP-Bypass-Techniques
- https://github.com/renniepak/CSPBypass

### Change type

- [x] `bugfix`
- [ ] `improvement`
- [ ] `feature`
- [ ] `api`
- [ ] `other`

### Release notes

- Restrict Github Gists to a more strict URL format.