Allow slurping of legacy multiplayer routes. (#5181)
When a logged in user visits a legacy multiplayer room (`/q/r/roomId`)
they can now click the `Copy to my files` button to slurp the
multiplayer room to a file that gets created.

This currently reuses all the existing assets - the files don't get
re-uploaded. We will need to re-upload them so that we will be able to
associate them with the user. Decided to skip this step for now and come
back to it once we have the assets/user/file associations ready.

### Change type

- [ ] `bugfix`
- [x] `improvement`
- [ ] `feature`
- [ ] `api`
- [ ] `other`

### Test plan

1. Create a legacy multiplayer room.
2. Login to the app and use the multiplayer room link, but prefix it
with `/q` (so something like `/q/r/roomId`
3. Click the `Copy to my files` button.
4. You should now see the same contents in the newly created file.