fix clipboard file upload (#5223) (#5225)
In #4771, we tweaked how we read from the clipboard on native paste
events. When we've written to the clipboard, the clipboard API lets us
read that data back more accurately. When copying from the users local
filesystem though, files currently aren't exposed through the clipboard
API. This tweaks our paste code to still go first to the clipboard API,
but if it doesn't find anything very interesting there, it falls back to
the event files instead.

Unfortunately, this is an inconsistency between pasting with the
keyboard, and pasting with the UI. We cannot currently access these
files at all through the UI, as that must use the clipboard api.

### Change type

- [x] `bugfix`

### Release notes

- Fix pasting files copied from the local filesystem with cmd-v