Added toast instead of throwing an error for the case when the amount… (#5201)
This PR solves the issue #5188. Throwing an error for the case when a
user is trying to upload more files than maxFilesAtOnce is replaced with
toast.

_(for testing purposes and screen recording, I reduced the max amount of
files)_

https://github.com/user-attachments/assets/b7468fd0-f14f-41c0-96d9-01f7326a0ecd

P.S.: I'm not sure about the correction of the text, so feel free to
notify me, and I'll change it :)

### Change type

- [x] `bugfix`
- [ ] `improvement`
- [ ] `feature`
- [ ] `api`
- [ ] `other`

### Release notes

- added toast instead of throwing an error for the case when the amount
of files is bigger than `maxFilesAtOnce`