Exports DX pass (#5114)
Over the last few weeks we've had a lot of requests on discord around
asset resolution, exports, and the two together. Some of the APIs here
have evolved and change independently of each other over time, so I
wanted to take a pass at making them make sense with each other a bit
more. There are a few things going on in this diff:

1. **BREAKING** The export/copy-as JSON option has been removed. I think
this was only ever there as a debug helper, and it's impossible to
actually make use of the JSON once copied (it's not the same as .tldr
json which confuses people).
2. `exportToBlob` is deprecated in favour of a new `Editor.toImage`
method. `exportToBlob` has been the canonical 'turn the canvas into an
image' helper for a while, but it has quite a weird looking signature
and isn't very discoverable.
3. the `copyAs` and `exportAs` helpers have had a couple of args merged
into the options bag for consistency.
4. The `jpeg` format has been removed from `copyAs`. This is technically
a breaking API change, but since it never actually worked anyway due to
browser limitations i think its fine.
5. SVG exports now resolve assets according to how they'll be used:
- if it's for an SVG, we still use the existing
`shouldResolveToOriginal` behaviour
- if it's for a bitmap export, we request an image downscaled according
to the size it will appear in that resulting bitmap.
6. Better reference docs for several APIs around this stuff.
7. **BREAKING** the `useImageOrVideoAsset` hook now requires passing in
`width`, instead of reading `shape.props.w`. This is so it can be used
with shapes other than our own. Whilst this is technically a breaking
change, this limitation means its unlikely that it was used with many
custom shapes in practice.
8. The clamping that we used to apply in to `steppedScreenScale` in
`resolveAssetUrl` has been moved to our own implementations of
`TLAssetStore`. This is so implementors can make their own decisions
about the range of scalings they might want to use.
9. The `steppedScreenScale` limit changed from 1/8 to 1/32. This is
because when testing with full res photos from a modern smartphone, we
were still downloading a 1000+px image in order to render it at a few
hundred px across (and also we don't pay anything for these right now).
Happy to change now/in the future if this doesn't seem right though.

### Change type

- [x] `api`

### Release notes

#### Breaking changes / user facing changes
- The copy/export as JSON option has been removed. Data copied/exported
from here could not be used anyway. If you need this in your app, look
into `Editor.getContentFromCurrentPage`.
- `useImageOrVideoAssetUrl` now expects a `width` parameter representing
the rendered width of the asset.
- `Editor.getSvgElement` and `Editor.getSvgString` will now export all
shapes on the current page instead of returning undefined when passed an
empty array of shape ids.

#### Product improvement
- When exporting to an image, image assets are now downloaded at a
resolution appropriate for how they will appear in the export.

#### API changes
- There's a new `Editor.toImage` method that makes creating an image
from your canvas easier. (`exportToBlob` is deprecated in favour of it)
- `SvgExportContext` now exposes the `scale` and `pixelRatio` options of
the current export
- `SvgExportContext` now has a `resolveAssetUrl` method to resolve an
asset at a resolution appropriate for the export.
- `copyAs(editor, ids, format, opts)` has been deprecated in favour of
`copyAs(editor, ids, opts)`.
- `exportAs(editor, ids, format, name, opts)` has been deprecated in
favour of `exportAs(editor, ids, opts)`