Asset uploads (#5218)
Changes:
* We now use the main sync worker for tldraw app asset uploads. For old
assets we still continue to use the assets worker.
* Image resize worker is now deployed as part of our deploy dotcom
action. It receives the multiplayer worker url. If the asset origin is
same as the multiplayer url (we are fetching tldraw app asset) it uses a
service binding to request the asset. This is to avoid some permissions
issues we encountered when the image worker fetched directly from the
multiplayer worker. The fetching for existing assets should remain the
same. I added `IMAGE_WORKER` env variable to different environments for
our github actions.
* We now associate assets with files. We do this in two ways: 1. when
users upload the files we immediately add the `fileId` to the asset's
`meta` property. We then also write an entry into a new postgres table
called `asset`. 2. in some cases we do this server side (duplicating a
room, copy pasting tldraw content, slurping legacy multiplayer rooms).
The way it works is that the server checks all the tldraw app assets and
makes sure their meta is associated to the correct file. If it is not it
re-uploads the file to the uploads bucket and it updates the asset. It
does this when persisting a file and also on restore.
* There are some public API changes listed that were necessary to make
this work. They are listed below.

### Change type

- [ ] `bugfix`
- [x] `improvement`
- [ ] `feature`
- [ ] `api`
- [ ] `other`

### Release notes

**Breaking change**

- `@tldraw/tlschema`: `TLAssetStore.upload` used to return just the
`src` of the uploaded asset. It now returns `{src: string, meta?:
JsonObject}`. The returned metadata will be added to the asset record
and thus allows the users to add some additional data to them when
uploading.
- `@tldraw/editor`: `Editor.uploadAsset` used to return
`Promise<string>` and now returns `Promise<{ src: string; meta?:
JsonObject }> `