Use a more specific return type for getCropBox
This uses the more specific generic type for the type property, sets the
x, y and props properties required since they're always returned, and
doesn't include the optional properties that are never returned that
TlShapePartial included.

I could have been even more specific by marking props.crop as required
since it's always set (now it's a union with null, since that's what
ShapeWithCrop has), but I figured it could be nullable so the type
doesn't have to change if the implementation is changed to use null in
some cases.