David/int 677 guests see slurp failure modal (#5280)
- fix slurping images from /preview sign in. (it used an abortController
bound to the lifecycle of the dialog, but it should have been bound to
the lifecycle of the editor)
- prevent guest users trying to slurp files they don't own
- (unrelated) i made the route switching more robust. I saw a few times
in staging that it would get out of sync with my logged in status
somehow. so now we just force a refresh if the auth status changes and
it doesn't match the preview flag.

### Change type

- [x] `other`