Revert "Fix for resizing snapshot bug (#5211)" (#5292)
This reverts commit 4e8dbf183d7ccf788c36de38abc982dec0870a6a.

@steveruizok you mighta missed my comment on your PR:
> I was gonna approve but actually, before you land, I'd like to leave
this PR open so that I can better test this Sentry improvements:
https://github.com/tldraw/tldraw/pull/5221
> I was looking at the events, and I think it's when people are
"fighting" over a shape but it's a guess that I'd like to confirm with
more context.

So, before we suppress this error, I want to see if I can find out more
with the extra context/tooling that was recently added.


### Change type

- [ ] `bugfix`
- [ ] `improvement`
- [ ] `feature`
- [ ] `api`
- [x] `other`