Use the uncropped width when requesting an image shape asset
This fixes an issue where the more you cropped an image, the lower scale
the image would be resolved to. However, when cropping the image is
still rendered at the same size, it just shows only a part of the image.
Therefore use the uncropped size when resolving the asset instead, so it
keeps the correct scale.

Fixes #5299