Use the uncropped width when requesting an image shape asset (#5300)
This fixes an issue where the more you cropped an image, the lower scale
the image would be resolved to. However, when cropping the image is
still rendered at the same size, it just shows only a part of the image.
Therefore use the uncropped size when resolving the asset instead, so it
keeps the correct scale.

Fixes #5299

### Change type

- [x] `bugfix`
- [ ] `improvement`
- [ ] `feature`
- [ ] `api`
- [ ] `other`

### Test plan

1. Have an asset store that scales assets
2. Create an image shape
3. Crop the image
4. Observe that the more you crop the image, the lower scale it gets

- [ ] Unit tests
- [ ] End to end tests

### Release notes

- Fixed a bug where a cropped image would use lower scaled assets the
more it was cropped.