Emit a before-event from Editor (#5319)
This will emit an event from Editor when dispatch is called that can be
used to handle events before they are handled by tldraw.

The specific thing we need this for is to update camera options on wheel
and pointer events before tldraw handles them in order to have a
different wheelBehavior based on if you are using a mouse or a touchpad
(which is determined from a heuristic based on the wheel and pointer
events).

Describe what your pull request does. If you can, add GIFs or images
showing the before and after of your change.

### Change type

- [ ] `bugfix`
- [ ] `improvement`
- [ ] `feature`
- [x] `api`
- [ ] `other`

### Release notes

- Emit a `before-event` from Editor for events before they are handled
by tldraw.