support react 19 (#5293)
This diff adds support for react 19. React 19 is backwards compatible
with react 18, excluding the removal of some deprecated APIs. React 18.3
(which we already use) warns when these APIs are used.

"adding support" here involves two things:
1. upgrading our depenencies to ones that specify react 19 in their peer
deps array. I just upgraded everything that had a non-breaking upgrade
to go to.
2. adding react 19 to our own list of peer dependencies

We want to make sure that we can still be used with react 18, so we're
not upgrading to react 19 internally because we don't want to make use
of any of the new APIs that aren't 18-compatible.

I've made the decision here not to do anything like running tests
against both react versions. I think that this would be more effort to
set up (and have more on-going cost on CI times etc) than it would be
worth, and I trust react's guarantees about react 19's backwards
compatibility and the warnings in 18.3.

### Change type

- [x] `improvement`

### Release notes

- tldraw now supports react 19

---------

Co-authored-by: Mime Čuvalo <mimecuvalo@gmail.com>
Co-authored-by: huppy-bot[bot] <128400622+huppy-bot[bot]@users.noreply.github.com>