Public preview opt-in (#5324)
Instead of a full-blown banner, I just added an item in the menu, and a
little dot to entice people to click. this is a common pattern when apps
add new stuff and I think it works pretty well here, especially since
the goal is just to make sure the backend can handle the scale before we
go full switchover.

![Kapture 2025-01-30 at 11 23
56](https://github.com/user-attachments/assets/507ee977-8e8d-4231-8df8-bfc34fee8017)

### Change type

- [x] `other`