Combined sign in/up for dotcom (#5325)
Clerk's API for this is that combined sign in/up mode is enabled when a
`signInUrl` is provided but a `signUpUrl` isn't. They sort of seem to
have an assumption that there's a page somewhere that does sign in,
which we sort of don't have (its not in the app because we use modals)
but sort of do (clerk creates one automatically at
https://accounts.tldraw.com/sign-in).

This diff specifies the `signInUrl` (enabling combined sign in/up mode)
and creates a placeholder sign in page at that path (just in case
someone somehow gets sent to it, although I'm not sure how/why this
would happen).

Clerk likes react-router to be searching for a wildcard under
`/sign-in`. We don't have wildcard support in our routing stuff at the
moment, but I don't think we'd need to construct those URLs ourself
anyway, so I've just dynamically configured the route for now.

### Change type
- [x] `other`