Fix vscode publishing. (#5327)
I guess we had to upgrade the `@types/vscode` to`^1.96.0` to get it
working with react 19?

Updated the engine as well. This means that users will only get the
update if they are using vs code with that engine or higher. I guess
people update their editors so they should be using close to the latest
one.

### Change type

- [x] `bugfix`

### Release notes

- Fix packaging of the vs code extension.