Move to server side slurping. (#5348)
This moves duplication and legacy room slurping to the backend:
* It removes the the file open mode logic (no more setting it when
navigating, no more sending search params to the backend).
* We now insert a file into the db with an additional column called
`createSource` which is in the form of `f/kZTr6Zdx6TClE6I4qExV4`. In
this case this means duplication of an existing app file. For legacy
files we'd similarly use `/r|ro|v|s/kZTr6Zdx6TClE6I4qExV4` to use that
existing legacy file as a starting point.
* Room DO then sees this column and reads the data from the appropriate
source and initializes the room.
* We now retry getting the file record from the db, because otherwise we
could still end up with a "Not found" issue since the navigation might
happen before the replicator tells the room DO about the file creation.
* Assets will get re-uploaded with our existing
`maybeAssociateFileAssets` logic, that makes sure all the assets are
associated with the file.

We had to do this because in some cases the file got created in the db,
then an file update was dispatched to the replicator (due to the trigger
we have on the file table), which incorrectly the set room DO (its
document info was set to a non app file without the creation params). If
this happened before the user navigation hit the worker it would mean
that duplication as well as slurping legacy files end up in a "Not
found" error, even though the file was correctly inserted into the db.

### Change type
- [x] `improvement`

### Release notes

- Move duplicating and legacy file slurping to the server.