Fix tldr file upload (#5350)
This PR uploads the assets separately from the room data json blobs, to
avoid OOM exceptions on the server. I also consolidated the upload
lifecycle management.

Unfortunately, the way it's currently set up, the images get duplicated
on our backend because there's no way to assign the file ownership in
postgres without the file row already existing, but making that happen
with good UX is tricky. I looked into resolving it a number of ways but
it was getting a bit yak shavey so I decided to ignore the duplication
issue for now.

I think the ideal UX would be:

1. as soon as you drop the .tldr file onto the canvas we extract all the
assets into indexeddb, upload the records to R2, and create a file row.
3. we gradually upload the assets in the background as soon as we know
the file rows are in the db.
4. the files appear in the sidebar immediately, with loading spinners to
indicate that they are still uploading. You can open them up and make
edits and stuff while the uploading happens in the background

which is what I tried to implement to begin with, but it was getting too
heavy and I want to move on so I fell back to this.

### Change type

- [x] `bugfix`