Add 'report a problem' dialog (#5359)
The current methods of giving product feedback are kinda rubbish. You
have to either know tldraw is a github project and have an account
there, or you need to dm steve on twitter, or you need to find the link
to discord somewhere and find the right channel and bug us in there.

This PR aims to reduce the friction to providing feedback by supplying a
text box people can use to send us a message on discord, as well as
offering links to our github and discord. If people start abusing the
textbox we can disable it or add some abuse prevention beyond the rate
limiting I added here. Hopefully that won't happen for a while.


https://github.com/user-attachments/assets/d0cae3ef-cba4-45e0-a8d5-03918b23f433


### Change type

- [x] `other`