Use tick for asset debounce (#5361)
When a user zooms in, all of the image or video shapes will run a
debounced function that attempts to resolve a new URL for the asset.
This is part of the LOD work done in #3684 /
https://github.com/tldraw/tldraw/pull/3827. One interesting issue is
that the debounce function itself is relatively expensive. With many _n_
shapes on the canvas, the browser is adding and removing at least _n_
global timers on each frame.

For example, here's me zooming in and out on a bunch of images. The 

<img width="1335" alt="image"
src="https://github.com/user-attachments/assets/0548318a-52fd-44a9-8cee-a5b7dd323d51"
/>

<img width="458" alt="image"
src="https://github.com/user-attachments/assets/cd82014b-e48d-4f82-92bd-54f9280ba797"
/>

To avoid this work, this PR hooks into the editor's tick event, which
fires every 16ms. Rather than debouncing for 500ms, we count the ticks
and add / remove listeners from the editor.

We make a request when the screen size of an image changes. This can
happen when a user resizes a change or when a user changes the zoom. We
could further improve the latter case by firing a "zoom ended" or
"camera settled" event from the editor and only making those requests
then; but we'd still need to "debounce" the resizing story.

### Change type

- [ ] `bugfix`
- [x] `improvement`
- [ ] `feature`
- [ ] `api`
- [ ] `other`

### Test plan

1. Make lots of images of any size.
2. Zoom in and out.
3. Check performance tab for smooth sailing.

### Release notes

- Improve zoom performance when many shapes are in a document.