Improve published file experience (#5371)
This is by no means a perfect solution, but it's a lot better than the
current situation for logged in users, which is bad mainly because it
shows the sidebar. Showing the sidebar for published files:

- fails to give the file owner a sense of how other people will see the
published version
- prevents the viewer from really _feeling_ how published files are
different
- puts the sidebar in a confusing navigational state where nothing is
selected

With this PR

1. Logged in users see the 'anonymous' view (no sidebar) but with a
share button instead of the sign in button.
2. Logged in users also get a 'copy to my files' affordance in the three
dots menu.
3. We replace the 'Save a copy' menu item (which opens the file picker)
with a 'Download file' menu item (which downloads straight to the
downloads folder).
4. Clicking the tldraw logo takes you back to the main app view

I think there's still a lot of room for improvement here, but this feels
like a quick win to land ahead of launch

### Change type

- [ ] `bugfix`
- [x] `improvement`
- [ ] `feature`
- [ ] `api`
- [ ] `other`