ShapeUtil.configure for shape options (#5399)
This diff tweaks the shape options API added in #5349 so that it's no
longer global. Instead, shape utils have a `.configure()` method, which
accepts a partial of their options and returns a new shape util with
those options applied.

We also tweak the behaviour of passing in custom shape utils to play
nicer with this: previously, if you passed in a shape util with type
"arrow", an error would be thrown complaining about duplicate shape
utils because the default one was already specified. Now, we won't try
to include the default arrow util because we can see you using a custom
one. If you pass two utils with the type "arrow", you'll still get an
error.

### Change type
- [x] `api`

### Release notes

- introduces shape options & `ShapeUtil.configure`, a utility for
passing options to a shape util
- moves (unreleased) noteShapeResizeMode to NoteShapeOptions.resizeMode
- If you pass tldraw a shape util with the same type as a default, it'll
now replace the default rather than crash
- **BREAKING** `options.maxDrawShapePoints` should now be specified with
`DrawShapeUtil.configure({maxPoints})` and
`HighlightShapeUtil.configure({maxPoints})`