Secret local file index (#5419)
So on tldraw.com if you're logged out your content will be saved only in
indexeddb. Then if you added at least one shape, when you log in we
slurp it into an app file. We then _don't_ delete the indexeddb entry in
case something went wrong with the slurping. This did indeed appear to
happen one time with one of our users. Luckily their data was under the
default persistenceKey and we were able to link them to
`https://tldraw.com/lf/tldraw_document_v3` or whatever it was.

This `/lf/:persistenceKey` route is a local-only route that loads a
tldraw document from indexeddb, and it will create a document if one
doesn't already exist for the given :persistenceKey.

This PR adds an index page under the `/lf` route, so that you can see a
list of all the tldraw documents in your indexeddb. This will be useful
to us in case we ever get someone who complains that something they drew
while logged out disappeared after they logged in, but it wasn't under
the default persistenceKey of `tldraw_document_v3` (we generate a new
persistenceKey every time we slurp a local file into the app). We can
just point them to `https://tldraw.com/lf` and let them do any recovery
they need.


### Change type

- [ ] `bugfix`
- [ ] `improvement`
- [ ] `feature`
- [ ] `api`
- [x] `other`